authentication/requirement.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
pyjwt
python-dotenv
